+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 - [2026-05-16 23:46] main.py
 - [2026-05-16 23:48] racer.py
 - [2026-05-17 10:03] skill-manifest.yaml
+- [2026-05-17 10:06] scripts/penetration-test.py